support named colors arguments cli
